Rename CesiumSandcastle files and components to OGC3dTiles
