+++ assetActions.config.js
@@ -4,16 +4,16 @@ import GeoJsonIo from './src/actions/assets/GeoJsonIo.js';
 import Geoparquet from './src/actions/assets/Geoparquet.js';
 import Potree from './src/actions/assets/Potree.js';
 import Protomaps from './src/actions/assets/Protomaps.js';
-import CesiumSandcastle from './src/actions/assets/CesiumSandcastle.js';
+import OGC3dTiles from './src/actions/assets/OGC3dTiles.js';
 import Viewer3D from './src/actions/assets/Viewer3D.js';
 
 export default {
-  CesiumSandcastle,
+  //Felt
+  OGC3dTiles,
   CopcViewer,
   GeoJsonIo,
   Geoparquet,
   Potree,
   Protomaps,
   Viewer3D,
-//Felt
 };
