+++ linkActions.config.js
@@ -1,9 +1,9 @@
 // import Felt from './src/actions/links/Felt.js';
-import CesiumSandcastle from './src/actions/links/CesiumSandcastle.js';
+import OGC3dTiles from './src/actions/links/OGC3dTiles.js';
 import Protomaps from './src/actions/links/Protomaps.js';
 
 export default {
     // Felt, 
-    CesiumSandcastle, 
+    OGC3dTiles, 
     Protomaps
 };
